alteracoes 03/03/26

+++ Projeto-SA-main/src/main/java/com/locacao/service/ClienteService.java
@@ -1,9 +1,10 @@
 package com.locacao.service;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import org.springframework.http.HttpStatus;
 
 import com.locacao.model.Cliente;
 import com.locacao.repository.ClienteRepository;
@@ -18,33 +19,37 @@ public class ClienteService {
         return clienteRepository.findAll();
     }
 
-    /*public Cliente buscarPorId(Integer id) {
-        return clienteRepository.findById(id).orElse(null);*/
-
-
-        public Cliente buscarPorId(Integer id) {
-    return clienteRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Cliente com ID " + id + " não encontrado."));
+    public Cliente buscarPorId(Integer id) {
+        return clienteRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(
+                        HttpStatus.NOT_FOUND, "Cliente com ID " + id + " não encontrado."));
     }
 
-    public Cliente salvar(Cliente cliente) {
+    public Cliente salvar(ClienteRequestDTO dto) {
+        Cliente cliente = new Cliente();
+        cliente.setNome(dto.nome());
+        cliente.setCpf(dto.cpf());
+        cliente.setEmail(dto.email());
+        cliente.setTelefone(dto.telefone());
+        cliente.setEndereco(dto.endereco());
         return clienteRepository.save(cliente);
     }
 
-    public Cliente atualizar(Integer id, Cliente clienteAtualizado) {
+    public Cliente atualizar(Integer id, ClienteRequestDTO dto) {
         Cliente cliente = buscarPorId(id);
-        if (cliente == null) return null;
-
-        
+        cliente.setNome(dto.nome());
+        cliente.setCpf(dto.cpf());
+        cliente.setEmail(dto.email());
+        cliente.setTelefone(dto.telefone());
+        cliente.setEndereco(dto.endereco());
         return clienteRepository.save(cliente);
     }
 
     public void deletar(Integer id) {
-    clienteRepository.findById(id)
-        .orElseThrow(() -> new RuntimeException("Cliente com ID " + id + " não encontrado."));
-
-    clienteRepository.deleteById(id);
-
+        if (!clienteRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cliente com ID " + id + " não encontrado.");
+        }
+        clienteRepository.deleteById(id);
     }
 }
